# optimize groundtruths and predictions creation