--- a/api/valor_api/backend/__init__.py
+++ b/api/valor_api/backend/__init__.py
@@ -1,6 +1,7 @@
 from .core import (
     create_dataset,
     create_groundtruth,
+    create_groundtruths,
     create_model,
     create_or_get_evaluations,
     create_prediction,
@@ -42,6 +43,7 @@ __all__ = [
     "create_dataset",
     "create_model",
     "create_groundtruth",
+    "create_groundtruths",
     "create_prediction",
     "delete_dataset",
     "delete_model",