--- a/api/valor_api/backend/core/datum.py
+++ b/api/valor_api/backend/core/datum.py
@@ -7,6 +7,29 @@ from valor_api.backend import models
 from valor_api.backend.query import Query
 
 
+def create_datums(
+    db: Session, datums: list[schemas.Datum], datasets: list[models.Dataset]
+) -> list[models.Datum]:
+    rows = [
+        models.Datum(
+            uid=datum.uid,
+            dataset_id=dataset.id,
+            meta=datum.metadata,
+        )
+        for datum, dataset in zip(datums, datasets)
+    ]
+
+    try:
+        db.add_all(rows)
+        db.commit()
+    except IntegrityError:
+        db.rollback()
+        # TODO: fix this exception
+        raise exceptions.DatumAlreadyExistsError("")
+
+    return rows
+
+
 def create_datum(
     db: Session,
     datum: schemas.Datum,
@@ -28,6 +51,11 @@ def create_datum(
     ----------
     models.Datum
         The datum.
+
+    Raises
+    ----------
+    exceptions.DatumAlreadyExistsError
+        If the datum already exists in the database.
     """
     try:
         row = models.Datum(