--- a/migrations/sql/00000007_add_indexes.down.sql
+++ b/migrations/sql/00000007_add_indexes.down.sql
@@ -1,0 +1,2 @@
+DROP INDEX idx_annotation_datum_id
+DROP INDEX idx_annotation_model_id