… api/tests/functional-tests/backend/core/test_annotation.py
… api/tests/functional-tests/backend/core/test_evaluation.py
… api/tests/functional-tests/backend/core/test_geometry.py
… api/tests/functional-tests/backend/core/test_label.py
… api/tests/functional-tests/backend/metrics/test_classification.py
… api/tests/functional-tests/backend/metrics/test_detection.py
… api/tests/functional-tests/backend/metrics/test_segmentation.py
… api/tests/functional-tests/backend/query/test_ops.py
… api/tests/functional-tests/conftest.py
… api/tests/functional-tests/crud/test_create_delete.py
… api/tests/functional-tests/crud/test_read.py
… api/tests/unit-tests/test_main.py
… api/valor_api/backend/core/prediction.py
… api/valor_api/backend/metrics/metric_utils.py
… api/valor_api/backend/models.py
… api/valor_api/crud/__init__.py
… api/valor_api/crud/_create.py
… api/valor_api/main.py
… migrations/sql/00000007_add_indexes.up.sql