This optimizes the adding of groundtruths and predictions by

1. actually add them in the batches they were sent instead of one by one
2. use bulk adds with mappings instead of ORM when possible
3. skip some unnecessary validation checks (specifically when adding groundtruth data, we do not need to both check if the datum does not exist and then check if annotations do not exist, since datum creation and groundtruth annotation creation must happen together).
4. Add two indexes on the `Annotation` table, one for `model_id` and one for `datum_id`.

This optimization was spurred by seeing slow performance on the demo instance. I tested by running the API locally but pointing to the postgres hosted on the demo instance. For the example data I was using running off the `main` branch gave the following speeds:

- groundtruth creation: 517s
- prediction creation: 531s

with the code changes except for the additional index (i.e. just 1-3 above) the times became

- groundtruth creation: 12s
- prediction creation: 40s

adding those two indexes on top of this, the times became

- groundtruth creation: 12s
- prediction creation: 25s